added delete features for posts and also a javascript modal features to confirm before delete

+++ admin/includes/post_view_all.php
@@ -52,7 +52,7 @@
                             echo "<td>$post_users</td>";
                             echo "<td><a href=''>View Post</a></td>";
                             echo "<td><a href='./posts.php?source=post_update&p_id={$post_id}'>Edit</a></td>";
-                            echo "<td><a href='./posts.php?source=post_delete&p_id={$post_id}'>Delete</a></td>";
+                            echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this Post')\" href='./posts.php?source=post_delete&p_id={$post_id}'>Delete</a></td>";
                         }
